+++ src/main/java/com/cxb/storehelperserver/service/HalfgoodService.java
@@ -37,6 +37,9 @@ public class HalfgoodService {
     private OriginalHalfgoodRepository originalHalfgoodRepository;
 
     @Resource
+    private OriginalRepository originalRepository;
+
+    @Resource
     private AttributeTemplateRepository attributeTemplateRepository;
 
     @Resource
@@ -125,6 +128,9 @@ public class HalfgoodService {
             return RestResult.fail(msg);
         }
 
+        if (!originalHalfgoodRepository.delete(halfgood.getGid(), halfgood.getId())) {
+            return RestResult.fail("删除半成品关联原料失败");
+        }
         if (!halfgoodAttrRepository.delete(halfgood.getId())) {
             return RestResult.fail("删除半成品属性失败");
         }
@@ -173,7 +179,8 @@ public class HalfgoodService {
             return RestResult.fail("获取公司信息异常");
         }
 
-        int total = halfgoodRepository.total(group.getGid(), search);
+        int gid = group.getGid();
+        int total = halfgoodRepository.total(gid, search);
         if (0 == total) {
             val data = new HashMap<String, Object>();
             data.put("total", 0);
@@ -181,7 +188,7 @@ public class HalfgoodService {
             return RestResult.ok(data);
         }
 
-        val commodities = halfgoodRepository.pagination(group.getGid(), page, limit, search);
+        val commodities = halfgoodRepository.pagination(gid, page, limit, search);
         if (null == commodities) {
             return RestResult.fail("获取半成品信息异常");
         }
@@ -205,6 +212,16 @@ public class HalfgoodService {
                 tmp.put("attrs", list);
                 for (THalfgoodAttr attr : attrs) {
                     list.add(attr.getValue());
+                }
+            }
+
+            // 关联来源
+            TOriginalHalfgood originalHalfgood = originalHalfgoodRepository.find(gid, c.getId());
+            if (null != originalHalfgood) {
+                TOriginal original = originalRepository.find(originalHalfgood.getOid());
+                if (null != original) {
+                    tmp.put("oid", original.getId());
+                    tmp.put("oname", original.getName());
                 }
             }
         }
